requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
email-validator==1.1.3
pymongo==4.3.3
dnspython==2.3.0
werkzeug==2.0.1